requirements/pytest.txt
dj_database_url~=2.0.0
pytest~=7.4.0
pytest-cov~=4.1.0
pytest-django~=4.5.2
pytest-dotenv~=0.5.2